+++ src/widget/timer_widget.rs
@@ -16,6 +16,7 @@
 use piston_window::*;
 use std::time::{SystemTime, UNIX_EPOCH};
 
+use crate::core::callbacks::CallbackEvent;
 use crate::core::point::*;
 use crate::widget::config::*;
 use crate::widget::widget::*;
@@ -32,6 +33,7 @@ pub struct TimerWidget {
     enabled: bool,
     initiated: u64,
     timeout: u64,
+    event: Option<CallbackEvent>,
 }
 
 /// Helper function that returns the current time in milliseconds since the `UNIX_EPOCH`.  This
@@ -60,23 +62,24 @@ impl TimerWidget {
             enabled: true,
             initiated: time_ms(),
             timeout: 0,
+            event: None,
         }
     }
 
-    //    // Called to check the time since initiation, and call the timeout function when a timer has
-    //    // been triggered.
-    //    fn tick(&mut self) {
-    //        if !self.enabled {
-    //            return;
-    //        }
-    //
-    //        let elapsed = time_ms() - self.initiated;
-    //
-    //        if elapsed > self.timeout {
-    //            self.initiated = time_ms();
-    //            self.timeout();
-    //        }
-    //    }
+    // Called to check the time since initiation, and call the timeout function when a timer has
+    // been triggered.
+    fn tick(&mut self) {
+        if !self.enabled {
+            return;
+        }
+
+        let elapsed = time_ms() - self.initiated;
+
+        if elapsed > self.timeout {
+            self.initiated = time_ms();
+            self.event = Some(CallbackEvent::TimerTriggered { widget_id: 0 });
+        }
+    }
 
     /// Enables or disables the timer.  When disabled, the timer will not initiate the callback
     /// function.  When re-enabled, the initiation time resets, so the timer will reset back to
@@ -85,21 +88,6 @@ impl TimerWidget {
         self.enabled = enabled;
         self.initiated = time_ms();
     }
-
-    //    /// Sets the closure function for the timer when a timeout has been triggered.  This closure
-    //    /// needs to be `Boxed`.
-    //    pub fn on_timeout(&mut self, callback: BlankCallback) {
-    //        self.callbacks()
-    //            .put(CALLBACK_TIMER, CallbackTypes::BlankCallback { callback });
-    //    }
-    //
-    //    /// Calls the timeout function.
-    //    fn timeout(&mut self) {
-    //        match self.callbacks().get(CALLBACK_TIMER) {
-    //            CallbackTypes::BlankCallback { callback } => callback(),
-    //            _ => (),
-    //        }
-    //    }
 
     /// Sets the timeout in milliseconds for this timer.  Will trigger a call to the function
     /// set in `on_timeout` when triggered, and will continue to call that function until this
@@ -131,9 +119,22 @@ impl Widget for TimerWidget {
         make_unsized()
     }
 
+    /// This function injects events, as a timeout event only occurs once.
+    fn injects_events(&mut self) -> bool {
+        true
+    }
+
+    /// Returns an injected event where appropriate.
+    fn inject_event(&mut self) -> Option<CallbackEvent> {
+        let return_obj = self.event.clone();
+
+        self.event = None;
+        return return_obj;
+    }
+
     /// Does not draw anything - only calls the timer `tick()` function to increment the
     /// timer.
     fn draw(&mut self, _context: Context, _graphics: &mut G2d, _clip: &DrawState) {
-        //        self.tick();
+        self.tick();
     }
 }